prettier refactoring
